refactor(FSD): move root layout to src/app/layouts
