+++ src/app/layouts/root/index.ts
@@ -1,0 +1,1 @@
+export { RootLayout } from './ui/layout';
